renamed mediumcard component
